added 4th spawn location

+++ Lords-of-Distortion/Assets/Scripts/LobbyInstanceManager.cs
@@ -120,7 +120,7 @@ public class LobbyInstanceManager : MonoBehaviour {
 	}
 
 	//this function should be called by the server arena manager.
-	public double SpawnPlayers(List<Vector3> spawnLocations){
+	public float SpawnPlayers(List<Vector3> spawnLocations){
 		Dictionary<NetworkPlayer, PlayerOptions>.KeyCollection players = playerOptions.Keys;
 		int i = 0;
